move find_config_file to config module
