+++ git-cliff-core/Cargo.toml
@@ -83,6 +83,7 @@ urlencoding = "2.1.3"
 cacache = { version = "13.1.0", features = ["mmap", "tokio-runtime"], default-features = false }
 time = "0.3.44"
 chrono = { version = "0.4.41", features = ["serde"] }
+tempfile = "3.27.0"
 
 [dependencies.git2]
 version = "0.20.3"
