+++ git-cliff/Cargo.toml
@@ -63,7 +63,6 @@ rand = { version = "0.9.2", optional = true }
 reqwest = { workspace = true, optional = true }
 url.workspace = true
 pathdiff = "0.2.3"
-tempfile = "3.27.0"
 
 [dependencies.git-cliff-core]
 version = "2.12.0" # managed by release.sh
